Move ReplayClient and ReplayClientContext to packages/shared
