+++ packages/bvaughn-architecture-demo/components/console/MessagesList.tsx
@@ -1,9 +1,10 @@
 import { Message } from "@replayio/protocol";
 import { useContext } from "react";
 
+import { ReplayClientContext } from "../../../shared/client/ReplayClientContext";
+
 import { ConsoleFiltersContext } from "../../src/contexts/ConsoleFiltersContext";
 import { FocusContext } from "../../src/contexts/FocusContext";
-import { ReplayClientContext } from "../../src/contexts/ReplayClientContext";
 import useFilteredMessages from "../../src/hooks/useFilteredMessages";
 import { getMessages } from "../../src/suspense/MessagesCache";
 import { getClosestPointForTime } from "../../src/suspense/PointsCache";
