+++ packages/bvaughn-architecture-demo/components/inspector/KeyValueRenderer.tsx
@@ -2,7 +2,8 @@ import { Object as ProtocolObject, PauseId, Value as ProtocolValue } from "@repl
 import classNames from "classnames";
 import { ReactNode, Suspense, useContext } from "react";
 
-import { ReplayClientContext } from "../../src/contexts/ReplayClientContext";
+import { ReplayClientContext } from "../../../shared/client/ReplayClientContext";
+
 import { getObjectWithPreview } from "../../src/suspense/ObjectPreviews";
 
 import Loader from "../Loader";
@@ -13,7 +14,6 @@ import styles from "./KeyValueRenderer.module.css";
 import PropertiesRenderer from "./PropertiesRenderer";
 import useClientValue from "./useClientValue";
 import ValueRenderer from "./ValueRenderer";
-import HTMLElementRenderer from "./values/HTMLElementRenderer";
 
 // Renders a protocol Object/ObjectPreview as a key+value pair.
 //
