+++ packages/bvaughn-architecture-demo/components/inspector/ValueRenderer.tsx
@@ -1,7 +1,8 @@
 import { PauseId, Value as ProtocolValue } from "@replayio/protocol";
 import { FC, memo, useContext } from "react";
 
-import { ReplayClientContext } from "../../src/contexts/ReplayClientContext";
+import { ReplayClientContext } from "../../../shared/client/ReplayClientContext";
+
 import { getObjectWithPreview } from "../../src/suspense/ObjectPreviews";
 
 import useClientValue from "./useClientValue";
