+++ packages/bvaughn-architecture-demo/components/inspector/values/HTMLElementRenderer.tsx
@@ -1,12 +1,13 @@
 import { PauseId, Value as ProtocolValue } from "@replayio/protocol";
 import { useContext } from "react";
 
-import { ReplayClientContext } from "../../../src/contexts/ReplayClientContext";
+import { ReplayClientContext } from "../../../../shared/client/ReplayClientContext";
+
 import { getObjectWithPreview } from "../../../src/suspense/ObjectPreviews";
 import { filterNonEnumerableProperties, Value as ClientValue } from "../../../src/utils/protocol";
 
 import useClientValue from "../useClientValue";
-import ValueRenderer from "../ValueRenderer";
+
 import ClientValueValueRenderer from "./ClientValueValueRenderer";
 
 import styles from "./shared.module.css";
