+++ packages/bvaughn-architecture-demo/tsconfig.json
@@ -18,6 +18,6 @@
       "components/*": ["components/*"]
     }
   },
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "../shared/**/*.tsx"],
   "exclude": ["node_modules"]
 }
