--- packages/markerikson-stack-client-prototype/src/contexts/ReplayClientContext.tsx
@@ -1,6 +1,0 @@
-import { createContext } from "react";
-import { replayClient, ReplayClientInterface } from "../client/ReplayClient";
-
-export type ReplayClientContextType = ReplayClientInterface;
-
-export const ReplayClientContext = createContext<ReplayClientContextType>(replayClient);
